==> analysis/0c_cleaning_for_1trial_back.Rmd <==
---
title: "3_stage_2_analyses"
author: "Paul Bloom"
date: "1/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(Matrix)
library(tidyverse)
library(lme4)
library(brms)
library(tidybayes)
library(emmeans)
library(easystats)
theme_set(theme_bw())


danl = read_csv('../clean_data/danl_space_treasure_comps.csv') %>%
  group_by(id) %>%
  dplyr::mutate(state_match = ifelse(state == lag(state), 'Same State', 'Different State'),
                next_match_button = lag(choice_2)==choice_1) %>%
  ungroup() %>%
  dplyr::rename('trial_num'='trial') %>%
  dplyr::filter(trial_num > 10) %>%
  mutate(cohort = 'danl')

hartley = read_csv('../clean_data/hartley_lag.csv') %>%
  group_by(id) %>%
  dplyr::mutate(state_match = ifelse(state == lag(state), 'Same State', 'Different State'),
                next_match_button = lag(choice_2)==choice_1) %>%
  dplyr::filter(age <= max(danl$age),
                age >= min(danl$age)) %>%
  ungroup()
                

simulated = load('../clean_data/sims_no_persev.rda')

sims = rbind(unnested_mb %>% dplyr::mutate(cohort = 'Simulated MB'),
             unnested_mf %>% dplyr::mutate(cohort = 'Simulated MF'))

sims = dplyr::select(sims, id=index, trial_num = trial, reward=rewardStage2, choice_1 = stage1Choice,
                     choice_2=choiceStage2, last_reward=prevReward, stay1=stay, state = stateStage2, transition=transitionStage2, cohort) %>%
  group_by(cohort, id) %>%
  mutate(choice_1 = choice_1 - 1,
         choice_2 = choice_2 - 1,
         stay2 = ifelse(choice_2 == lag(choice_2, n = 1), 1, 0),
         last_transition = lag(transition, n=1),
         last_reward = lag(reward, n = 1),
         last_choice_1 = lag(choice_1, n=1),
         last_choice_2 = lag(choice_2, n = 1),
         next_match_button = lag(choice_2)==choice_1,
         state_match = ifelse(lag(state, n=1)==state, "Same State", "Different State")) %>%
  ungroup() %>%
  dplyr::filter(trial_num >10 ) 


danl = dplyr::filter(danl, age > min(hartley$age))

# Create summary columns
danl = danl %>%
  group_by(id) %>%
  dplyr::mutate(total_points = sum(reward)) %>%
  ungroup() %>%
  mutate(total_points = scale(total_points))

hartley = hartley %>%
  group_by(id, cohort) %>%
  dplyr::mutate(total_points = sum(reward)/n()) %>%
  ungroup() %>%
  mutate(total_points = scale(total_points))

all_data = plyr::rbind.fill(hartley, danl) %>%
  plyr::rbind.fill(., sims) %>%
  dplyr::group_by(cohort, id) %>%
  dplyr::mutate(rt_2_z = scale(rt_2, center = TRUE, scale = TRUE)) %>%
  dplyr::ungroup()

```


```{r}
combined_data_for_models = all_data %>%
  dplyr::select(id, cohort, trial_num, choice_1, transition, choice_2, reward, last_reward, last_choice_1, last_choice_2, last_transition,
                next_match_button, state_match, stay1, stay2)

save(all_data, combined_data_for_models, file = '../clean_data/combined_data_for_models.rda')

```
